requirements.txt
Flask==2.0.3
paho-mqtt==1.6.1
pandas==1.3.5
python-configuration==0.8.2
python-dotenv==0.19.2
gevent==21.12.0